seonize-backend/requirements.txt
# Seonize Backend Dependencies

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Client
httpx>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver

# Cache
redis>=5.0.0

# AI APIs
google-generativeai>=0.8.0

# Web Scraping
beautifulsoup4>=4.12.0

# NLP
jieba>=0.42.0
scikit-learn>=1.5.0

# Utilities
python-dotenv>=1.0.0
